DxLatamDjPortal/requirements.txt
aioredis==1.3.1
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
Brotli==1.0.9
cffi==1.14.5
channels==3.0.3
click==7.1.2
constantly==15.1.0
cryptography==3.4.6
daphne==3.0.1
dash==1.19.0
dash-core-components==1.15.0
dash-html-components==1.1.2
dash-renderer==1.9.0
dash-table==4.11.2
Django==2.2.19
django-composite-foreignkey==1.1.0
django-crispy-forms==1.11.2
django-csv==1.2
django-currentuser==0.5.3
django-filter==2.4.0
django-mssql-backend==2.8.1
django-pandas==0.6.4
django-phonenumber-field==5.1.0
django-plotly-dash==1.6.2
django-redis==4.12.1
django-widget-tweaks==1.4.8
dpd-components==0.1.0
Flask==1.1.2
Flask-Compress==1.9.0
future==0.18.2
hiredis==1.1.0
hyperlink==21.0.0
idna==3.1
incremental==21.3.0
install==1.3.4
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
msgpack==1.0.2
numpy==1.20.1
pandas==1.2.3
phonenumbers==8.12.21
pip==21.1.2
plotly==4.14.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyodbc==4.0.30
pyOpenSSL==20.0.1
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
redis==3.5.3
retrying==1.3.3
service-identity==18.1.0
setuptools==47.1.0
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
Twisted==21.2.0
txaio==21.2.1
Werkzeug==1.0.1
zope.interface==5.3.0
